+++ packages/backend/src/modules/user/infra/knex/repositories/UserRepository.ts
@@ -1,22 +1,21 @@
 import { ICreateUser } from '@modules/user/dtos/ICreateUserDTO';
 import { IUpdateUserDTO } from '@modules/user/dtos/IUpdateUserDTO';
-import { BaseRepository } from '@utils/BaseRepository';
 
 import { KnexEntity, knexQuery } from '@shared/infra/knex/knexfile';
 
 import IUsersRepository from '../../repositories/IUsersRepository';
 import { User } from '../entities/User';
 
-export class UserRepository extends BaseRepository implements IUsersRepository {
+export class UserRepository implements IUsersRepository {
   public async create(user: ICreateUser): Promise<string | null> {
     const createUser = (await knexQuery()
       .insert(user)
       .into('users')
       .then(() =>
         knexQuery().select('user_id').from('users').where('email', user.email),
-      )) as [{ userId: string }];
+      )) as [{ user_id: string }];
 
-    return createUser.shift()?.userId ?? null;
+    return createUser.shift()?.user_id ?? null;
   }
 
   public async findByEmail(email: string): Promise<KnexEntity<User> | null> {
@@ -29,19 +28,17 @@ export class UserRepository extends BaseRepository implements IUsersRepository {
   }
 
   async findByUserId(userId: string): Promise<KnexEntity<User> | undefined> {
-    const id = this.setLocalId(userId);
     const user = await knexQuery()
       .select()
       .table('users')
-      .where({ id })
+      .where({ id: userId })
       .then((response) => response[0]);
 
-    return { ...user, id: this.setGlobalId(user.id) };
+    return user;
   }
 
   async delete(userId: string): Promise<void> {
-    const id = this.setLocalId(userId);
-    await knexQuery().delete().table('users').where({ id });
+    await knexQuery().delete().table('users').where({ id: userId });
   }
 
   async update(user: IUpdateUserDTO): Promise<void> {
